Add query subcommand
