+++ Cargo.toml
@@ -17,6 +17,7 @@ camino = "1.1.6"
 clap = { version = "4.5.4", features = ["derive", "wrap_help", "env"] }
 clap_complete = "4.5.1"
 clap_mangen = { version = "0.2.20", optional = true }
+comfy-table = "7.1.0"
 command-error = { version = "0.4.0", features = [ "tracing" ] }
 fs-err = "2.11.0"
 itertools = "0.12.1"
