+++ src/author.rs
@@ -3,16 +3,7 @@
 #[serde(rename_all = "camelCase")]
 #[allow(dead_code)]
 pub struct Author {
-    email: String,
-    #[serde(flatten)]
-    user: User,
-}
-
-/// A Gerrit user.
-#[derive(serde::Deserialize, Debug)]
-#[serde(rename_all = "camelCase")]
-#[allow(dead_code)]
-pub struct User {
-    name: String,
-    username: String,
+    pub email: Option<String>,
+    pub name: String,
+    pub username: String,
 }
